test: fix import errors in benchmarks
