+++ tests/benchmarks/test_pydantic.py
@@ -1,9 +1,36 @@
-from .models import PydanticUser
 import pytest
+
+pytest.importorskip("pydantic", reason="Pydantic adapter requires 'pydantic' package")
+from typing import Literal
+
+from pydantic import BaseModel, Field
 
 from conformly import case, cases
 from conformly.parsing.adapters.pydantic import parse
 from conformly.resolver.resolve import resolve_model
+
+
+class PydanticAddress(BaseModel):
+    street: str = Field(..., min_length=5, max_length=50)
+    city: str = Field(..., min_length=5, max_length=50)
+    country: Literal["Germany", "Russia", "China", "USA"]
+
+
+class PydanticProfile(BaseModel):
+    bio: str | None = Field(None, max_length=300)
+    rating: float = Field(..., gt=0.0, le=100.0)
+    role: Literal["admin", "moderator", "user", "guest"]
+
+
+class PydanticUser(BaseModel):
+    id: int
+    username: str = Field(..., min_length=3, max_length=25)
+    email: str = Field(..., pattern=r"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$")
+    phone: str = Field(..., pattern=r"^\+?[1-9]\d{1,14}$")
+    age: int = Field(..., ge=18, le=120)
+    address: PydanticAddress
+    profile: PydanticProfile
+    is_active: bool = True
 
 
 @pytest.mark.benchmark
